Workaround: Create groups for webrtc and socketio headers.
We can't install cocoapods for the respokeSDK target since respokeSDK is
a static library. This is a cocoapods limitation. Instead, we install cocoapods
for the respokeSDKTests target. This makes the cocoapod public headers accessible
to the respokeSDKTests sources, but not the respokeSDK sources.

As a result, we have to manually create groups for the headers so that the
respokeSDK sources can include them.