Fix random crash related to PeerConnectionFactory
Previously, the tests were sporadically crashing with an
EXC_BAD_ACCESS exception pointing at the RTCMediaStreamTrack dealloc
of the webrtc library. This was an extremely cryptic problem that
could have been causing a myriad of subtle difficult-to-understand
bugs. With this change, the PeerConnectionFactory now is only
initialized and instantiated once per application load. Bazinga!