Upgrade cocoapods version for xcode 7 support
When installing the pods using `pod install` with an older version
of cocoapods, it was exploding with an ASSERTION_ERROR. Upgrading
cocoapods fixed it, and subsequently it changed this value in the .lock
file. I don't know if it matters that this value changed, but I thought
I should check it in just to be sure.