Add RespokeVersion.h/.m for rudimentary versioning
This patch adds a very simplistic method of obtaining the version
information from inside the SDK. This was needed for proper version
reporting in the sdk header. In the future we'll probably want to
automate the updating of this file or implement some alternative
method of allowing the SDK access to its version, but for now this
works.