Add group message history support
This patch adds support for interacting with the Respoke group message
history APIs and flagging a message to be persisted to history.

A new method `joinWithSuccessHandler` has been added to RespokeGroup
to allow it to be joined when it is not currently joined. This was
needed because when retrieving group message history prior to joining
that group, there is no group in the client's group cache to associate
the message with. When creating the group, it did not make sense to
automatically subscribe the user to the group, so the group is created
and added to the cache without subscribing it in Respoke.

A new `persist` parameter has been added to RespokeEndpoint's
`sendMessage` message. This parameter allows a message to be flagged
for inclusion in the message history for the group. At this time,
this flag is only effective if the account being used to connect to
Respoke has had the message history feature enabled. Respoke message
history is currently an alpha quality feature.

Two new methods have been added to RespokeClient to allow retrieval
of group message history.

- `getGroupHistoriesForGroupIDs` allows the user to retrieve history
for multiple groups simultaneously. By default, the maximum number of
messages is set to 1, but the `maxMessages` parameter may be specified
to retrieve more messages from the history for each group specified.
When multiple messages are pulled from the history, the messages are
ordered by timestamp in reverse-chronological order.

- `getGroupHistoryForGroupID` allows the user to retrieve history
for a single group. By default the maximum number of messages is set
to 50, but the `maxMessages` parameter may be specified to retrieve
more messages from the history. Also, the messages that are retrieved
are by default the most recent messages available. To pull older
messages, specify the `before` param to indicate that you want to
pull messages older than that NSDate. The messages are ordered by
timestamp in reverse-chronological order.